Batch 2 - 60 files
The 2nd (and final) batch of files for this directory, containing a second and final set of 60 image files in JPG format.
I spent a lot of time re-uploading all these images. I am finally done for the day, however, about 4 months of content is still missing and needs to be found. I am caught up with what I have for now. Updates will continue to be casual again tomorrow, until I take the time to go and look for the missing 4 months of images that are on my external hard drives.