Batch 3 - 1 sub-directory (100 image files) (final batch)
The 3rd (and final) batch of files (out of 3) for this directory, containing the 3rd and final sub-directory, with 100 image files in JPG format. I can't believe I got an exact multiple of 100 of pictures. I was just thinking the other day that it would hardly ever happen, then it happens 2 days later. Weird accomplishment.